requirements.txt
asgiref==3.3.4
decouple==0.0.7
Django==3.2.3
django-filter==2.4.0
djangorestframework==3.12.4
install==1.3.4
pytz==2021.1
sqlparse==0.4.1
